working with reducers and reducers imports of act
