+++ src/redux/dialogs-reducer.js
@@ -18,4 +18,11 @@ const dialogsReducer = (state, action) => {
   return state;
 };
 
+export const sendMessageCreator = () => ({ type: SEND_MESSAGE });
+
+export const updateNewMessageBodyCreator = (body) => ({
+  type: UPDATE_NEW_MESSAGE_TEXT,
+  body: body,
+});
+
 export default dialogsReducer;
