+++ src/redux/profile-reducer.js
@@ -18,7 +18,15 @@ const profileReducer = (state, action) => {
       }
       break;
   }
+
   return state;
 };
+
+export const addPostActionCreator = () => ({ type: ADD_POST });
+
+export const updateNewPostTextActionCreator = (text) => ({
+  type: UPDATE_NEW_POST_TEXT,
+  newPostText: text,
+});
 
 export default profileReducer;
