+++ src/redux/state.js
@@ -1,10 +1,9 @@
 import profileReducer from "./profile-reducer";
 import dialogsReducer from "./dialogs-reducer";
-
-const ADD_POST = "ADD-POST";
-const UPDATE_NEW_POST_TEXT = "UPDATE-NEW-POST-TEXT";
-const UPDATE_NEW_MESSAGE_TEXT = "UPDATE_NEW_MESSAGE_TEXT";
-const SEND_MESSAGE = "SEND_MESSAGE";
+import {
+  addPostActionCreator,
+  updateNewPostTextActionCreator,
+} from "./profile-reducer";
 
 let store = {
   _state: {
@@ -96,20 +95,6 @@ let store = {
     this._callSubscriber(this._state);
   },
 };
-
-export const addPostActionCreator = () => ({ type: ADD_POST });
-
-export const updateNewPostTextActionCreator = (text) => ({
-  type: UPDATE_NEW_POST_TEXT,
-  newPostText: text,
-});
-
-export const sendMessageCreator = () => ({ type: SEND_MESSAGE });
-
-export const updateNewMessageBodyCreator = (body) => ({
-  type: UPDATE_NEW_MESSAGE_TEXT,
-  body: body,
-});
 
 export default store;
 window.store = store;
